requirements.txt
affine==2.4.0
asgiref==3.9.1
attrs==25.3.0
basemap==2.0.0
basemap_data==2.0.0
branca==0.8.1
Cartopy==0.24.1
cattrs==25.1.1
certifi==2025.7.14
cftime==1.6.4.post1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
fiona==1.10.1
flatbuffers==25.2.10
folium==0.20.0
fonttools==4.59.0
geopandas==1.1.1
h11==0.16.0
idna==3.10
inflection==0.5.1
jh2==5.0.9
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
netCDF4==1.7.2
niquests==3.14.1
numpy==2.3.2
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
PyJWT==2.10.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qh3==1.5.3
rasterio==1.4.3
rasterstats==0.20.0
referencing==0.36.2
requests==2.32.4
requests-cache==1.2.1
retry-requests==2.0.0
rioxarray==0.19.0
rpds-py==0.26.0
scipy==1.16.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.5.0
urllib3-future==2.13.900
wassima==1.2.2
xarray==2025.7.1
xyzservices==2025.4.0
